requirements.txt
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
configparser==7.2.0
contourpy==1.3.3
cycler==0.12.1
DateTime==5.5
deepdiff==8.6.1
dipy==1.11.0
fonttools==4.60.0
h5py==3.14.0
idna==3.10
joblib==1.5.2
kaggle==1.7.4.5
kiwisolver==1.4.9
matplotlib==3.10.0
nibabel==5.3.2
numpy==2.3.3
orderly-set==5.5.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
protobuf==6.32.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
setuptools-scm==9.2.0
six==1.17.0
text-unidecode==1.3
threadpoolctl==3.6.0
TIME-python==0.0.17
tqdm==4.67.1
trx-python==0.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
zope.interface==8.0
xgboost==3.0.5
pyarrow==19.0.0
webdriver-manager==4.0.2
zendriver==0.14.2
psutil==5.9.0

